Los Mochos and Puebla tile fixes (#1734)
* 18MEX: Fix Los Mochos hex

Los Mochos green is an L tile, while for yellow normal curved cities are
used. So special handling in upgrade needed. And also to show Los Mochos
L tile as a potential tile upgrade for tile 5 and 6 in the tile manifest.

* 18MEX: Auto lay Puebla tile when Mexico City is upgraded

When Mexico City is upgraded the action automatically triggers
an upgrade of the Puebla hex. The matching tiles for Puebla
has the same name, but P instead of MC.

To stop anyone upgrading Puebla directly, and game error check
is done.

Co-authored-by: Toby Mao <toby.mao@gmail.com>