18MEX: Correct mail income
Mail should only be given once per OR per corporation.

Minor change of having the log message for placing home token come
after the message that player X is operating the corporation.
This also makes it possible for the mail feature to be added
to the overloaded base method for placing home token.