Sort corporations in stock rounds.
For 1817 the amount of corporations gets confusing

Sort those that have been IPOed by their running order
Then others leave in symbol order

Done in base class to game so that other games (18CO and 1849 etc) can override.

This does have the slight impact to the UI that the position can change dynamically,
but this makes it clearer the change in stock price.