Correct price for 5% shares (on the server side)
share.rb and share_bundle.rb gave 0 as price for share counts
not evenly divisable by 10. This caused the 5% of NdM in 18MEX
to be excluded from the value (on the server side).

When paying out revenue, use float for num_shares to handle
the NdM case.

In both cases ceil is used to round up.