Improvements for handling merged corporation (#2005)
* Improvements for handling merged corporation

Merged corporation is now kept as entity for operating round,
with (CLOSED) after it. If current entity is merged, any remaining
actions is skipped silently.

The merged (closed) corporation are now skipped on entities
and game (stock) page, but remains on the spread sheet tab.

Correct removal of shares from merged corporation in 18MEX.

This fixes the following titles:
- Fixes #2003, the crash was caused by removing of a corporation
  from round entities - caused index error. Now the corporation
  remains but it will not act in the current operation round.
- Fixes #1913 for 18MEX - merged corporation does now remain on
  spread sheet, with all 10 shares in IPO, no share or par price,
  but do show operating history.
  For 1846, 18LA, 1817, Others? it depends on how close of
  corporation is handled.

Added close! and closed? methods to minor and corporation. These
only set closed attribute, sets owner to bank, and set floated to
false. Any clean up needed have to be done in the title that calls
the close! method.

* Corrections after review comments

* Further corrections after review comments