18MS: Remove extra 4+ train (fixes #2106) (#2110)
The 18MS implementation had erronously 4 4+ trains, instead of the
3 4+ trains according to the rules.

This removes the extra 4+, but this breaks any 18MS that has passed
4+ trains. It is not possible to fix it as it affect quite a lot
so I guess we will have to accept this and pin the games that break?

The spec fixture file was not possible to fix. So I remove it, and
push this PR and will use one of the completed games after this has been
deployed as a new spec. That will be added in the future.

I think it is best to wait with any other corrections of 18MS
until such a spec fixture file is added.