In case different percentages is in market, separate buy buttons
This is currently only for 18MEX but it is general code that will
be used as soon as there are different percentage shares in the
share pool.

In case there is just one type of shares the behavior will not be changed.

In case there are more the one type of percentage in the share pool
the percentage of the market share will be shown on the Buy button.
This is also presented even if only one of the types are buyable.

So, using 18MEX as an example:

* If 5% and 10% are in market pool, and player can buy both
  Show: "Buy 5% Market Share" "Buy 10% Market Share"

* If 5% and 10% are in market pool, but player can only buy 5%
  Show: "Buy 5% Market Share"

* If just 5% or 10% shares are in market pool
  Show "Buy Market Share"

* And this means - if not 18MEX
  Show "Buy Market Share"

This fixes #2097