18AL: Implement optional rules (fixes #2036) (#2139)
Optional rule 7a means that the first OR for any corporation, it
may lay 2 yellow tiles instead of 1 yellow or 1 upgrade.

Optional rule 7b means that the home location for L&N corporation
is moved to Decatur. This also means that hex A3 is turned into
an off board hex.

Optional rule 7c means that the number of 4D trains are unlimited.
This will change the number of 4D from 5 to 8. Theoretically
each of the 6 corporations may own 2 trains each, and there are
4 other permanent trains, so 8 4D is the max number of trains
that can be owned by corporations.

Optional rule 7d means that the 4T trains now are rusted by 7T
instead of phased out.