In case different percentages is in market, separate buttons (#2128)
This is currently only for 18MEX but it is general code that will
be used as soon as there are different percentage shares in the
share pool.

In case there is just one type of shares the behavior will not be changed.

In case there are more the one type of percentage in the share pool
the percentage of the market share will be shown on the Buy/Exchange button.
This is also presented even if only one of the types are buyable.

So, using 18MEX, and Buy, as an example:

* If 5% and 10% are in market pool, and player can buy both
  Show: "Buy 5% Market Share" "Buy 10% Market Share"

* If 5% and 10% are in market pool, but player can only buy 5%
  Show: "Buy 5% Market Share"

* If just 5% or 10% shares are in market pool
  Show "Buy Market Share"

* And this means - if not 18MEX, but still Buy
  Show "Buy Market Share"

* For Exchange a similar behavior

This fixes #2097