Remove reserved shares line from upper part of charter (fixes #2103)
Before change the upper part of the corporation charter could show:
  IPO
  (Reserved)
  Market
but after the change the (Reserved) line is removed.
Note: (Reserved) is "Trade-in" for 18MEX, but this is general.

The reserved shares are still shown further down.
For example, the NdM chater in 18MEX can show
     Trade-in shares: 5%, 5%, 10%

Note: The reserved shares are still deducted from the IPO count, so
it will just show available shares for buy.