18MEX: Temporarily remove it from new games creation
To avoid some frustration stop new games of 18MEX to be created.
That is because I plan to implement a couple of game breaking
things, and then it is good if as few ongoing games as possible
have to be deleted.

The alpha testing will resume as soon as those game-breaking changes
have been deployed.