18MEX: Introduce swap functionality (fixes #1921)
In 18MEX it is possible to swap a 5% share as part of a sell/buy.

If selling shares and there is a 5% share in the market, player
can choose to get the 5% as part of the deal, reducing the money
received for the sell. Note: Selling 30%, with a 5% swap, just
decreases stock market price as if 20% had been sold.

If buying shares and the player owns one 5% share, the player
can choose to use the 5% as a partial payment, reducing the
money needed by the value of the 5% share. This is counted
as a buy for all purposes.

Have introduced swap_sell and swap_share methods, which if
implemented will enable the swap buttons to appear in the GUI.
So far only 18MEX has these implemented.

Note! Swap button is Sell/Buy buttons, which do show info
about 5% share as part of deal.

A bunch of smaller bug fixes for 18MEX 5% share handling.
 - Now handles the 60% roof correctly for NdM.
   Note! If holding 55%, buying 10% is allowed if you swap
   your 5%. Similar if 25% in market, you can sell 30% if
   you swap in 5%.
 - Use SELL_MOVEMENT :down_by_10 instead of override of
   sell_shares_and_change_price
 - Override num_certs to handle the special case for NdM 5%.
   Have rewritten can_gain? to use similar guard as
   can_buy? and can_sell? (stop NdM actions before phase
   minors have merged)