Fixes for NdM 5% in market pool (#2234)
* 18MEX: Allow 5% NdM share buy even if at cert limit (fixes #2214)

It is allowed to buy 5% NdM shares even if at certificate limit.
Still cannot buy above 60% though.

* Improve "Buy Market Share" in case non-10% share

If the only share in the share pool is a non-10% share then show
the percentage for the share.

Example: 18MEX with just a 5% share in the share pool. This is
shown as "Buy 5% Market Share".