18MEX: Skip minor track lay if it cannot afford to lay in home hex (#2240)
If minor corporation does not have enough cash to lay a yellow in
its home hex, the track lay is skipped.

In case home hex has a yellow/green tile then normal track laying
check is done. This might mean that minor gets to manually pass
as there are no allowed track lays it can afford.

But that last is an exceptional case.