18TN: Fix "company buy during OR1" bug (fixes #1887)
Buy remembering all buys during a round, it is possible to do
a correct check for private buys during OR 1.

Reuse the company_seller round attribute, but change it to a
list instead of a nil/player attribute. This ment refactoring
its use in 1889 a bit. Renamed it to company_sellers to reflect
its changed use.
I do believe it can just contain one value in 1889 so the effect
should be the same.

Also added the buy-private-from-other-player feature that is
part of 18TN. This requires migration of 18TN as it adds
extra passes during SRs.