18TN: Civil War should affect triggering company (fixes #1873)
There was a bug if the triggering company's only train was the
one triggering the Civil War. In which case trains was still
empty when checking for Civil War token.

Have now included the trigger in the event even if it has no
other trains.

Also added a log line per corporation describing if it gets
Civil War token or not, to make it more visible.