Fix problems with Buy and Sell button changes
Previous commit introduced possibility to show non-standard
percentage in 18MEX, but it introduced problems in other
titles. So this commit improves the solution and solves
a couple of issues.

Disregard president shares for pool shares if there are other
shares that can be bought. This means it will not be possible
to buy president's share directly from the market unless it
is the last one. (And the title need to handle that case.)

Do not let Treasury or Short share presence affect GUI
presentation. Before this fix, the "Buy a 10% Market share"
appeared when Treasury share was present, instead of just
"Buy Market Share".

Also, do not use 10% as the standard share size. Instead use
the share_percent of the corporation in question. This solves
the problem of 20% appearing for 5 share corporations in 1817.

Similar fixes for Sell view. Problem reported in issue #2325.

Fixes #2316 and #2325