Fix consecutively operating receivership corporations
This problem was first encountered in game 16134; this commit also adds a
fixture game that is based on 16134, but includes the B&O route, and then is
ended early.

The B&O running its route and withholding drops its stock price, but when the
actual game was played, the bug skipped the B&O and didn't change its stock
price. With the route added in after the fact, the game breaks due to different
operating order.