18MEX: Fix sorting problem with minors in spreadsheet
If trying to sort the player columns, the minors throw an exception.
So I have added dummy implementations of num_shares_of and share_percent
for minors so that they can be sorted as well. They will end up at
top or bottom when sorted on shares for a player.

Fixes #2368 and #2370