18MEX: Fix bugs for valuation of share holdings (fixes #2355)
Valuation of NdM holdings was incorrect - if holding both 5% share(s)
and normal shares, the value of it was just the 5% share(s).

Have change the general code for bankruptcy sell - now it continues
to sell in case there are sellable bundles after the largest
package has been sold. And it will sell the biggest bundles first.
This means that the 5% shares of NdM in 18MEX will eventually be
sold.

Set 18MEX to decrease stock price per 10% sold - this means that
bankruptcy sell of one 5% at a time wont affect stock price.
Also ment that the overriden method that handle sell and change
stock price could be removed from 18MEX base class.

Refactored base class a bit to make it possible for 18MEX to
override parts of the base class to handle the 5% shares.
Also introduced all_bundles_for_corporation as it need to
be called from 18MEX. The bundles_for_corporation is overridden
by e.g. 18MEX but 18MEX also need to call the base method
directly when it needs to get bundles that include any and all
5% shares. (During valuation and dump.)