18MEX: Close KCMO private when its ability is used
This is an official errata to 18MEX that was published in
May 2020; it has been an error in the rules for MANY years
although the components in some version has the correct
rule.

Have updated the description to KCMO to reflect this.

Also remove CC label to avoid it cluttering up the hex.
(Actually this was as intended in the implementation but
 action.hex.tile.name were not the expected value...)

This removes some income to the owning corporation which might affect
some games. For e.g the fixture that was checked in for 18MEX.
The extra $20 that now disappear means that the forced
train purchase deduct $20 from one of the players which also
meant that TM is not sold out, meaning there is a shift in
operating order... And the final scores are adjusted, affecting
the final ranks...

So, I think game 13315 can be pinned to the previous version.
Have not check if any other 18MEX games are affected by this
change but it is likely. They can be pinned as well - fixing
them is not possible to do in an automated way.