18MEX: Correct cash at start (fixes #2447)
When minor was set up $100 was added to the minor so that it could
buy a 2 train. This meant that 3 x $100 was added to the game and
that the total bank was $300 to high.

Have removed the extra $100 for each minor and instead bought the
train for free.

This change broke the fixture as its bank had less than $300 left
in the bank after the penultimate OR (6.1), meaning OR 6.2 was
never intended to be run.
If other games break it should be for the same reason - an extra
OR (and possibly SR).