18MEX: Fix bug with delayed minor close (fixes #2457)
If the game uses the optional rule of delayed close of minors
the minor close in the Stock Round following the OR set where
the 5th 3 train was bought.

But the bug was that it was attempted to do the delayed minor
close in the Stock Round after that. So the delayed minor close
just had to be reset.