Minor fixes/refactoring
18MEX: Correct bank -> @bank in game class

corporation.rb: Make capitilization modifiable. This is for 1856
(and 18SJ) so that they can change capitilization for corporation
due to phase change.

corporation.rb, minor.rb: Extract the transfer part to a module
that can be shared. Today only corporations can transfer things
(because of merge) but there are titles that have minors that
can own things  which can be merged, so this is a preparation for that.