18CO Remove City Resv, Close Corps, Re-reserve Parred Corps
From phase 5 onward, cities are no longer reserved.
If a corporation can never place its home token, it is removed from the game.
Once a corporation is parred, its reservation is returned to the hex.
A hex might be a single city, but the corporation may upgrade it to place its home token.