Fix all_bundles_for_corporation sorting when same percentage share exists (#2731)
The assumption was made that the President's certificate would always have a higher percentage and thus price than any other share. However, in a game where there are shares with the same percentage, the President's cert could wind up anywhere in the shares list. This could cause any or all bundles to contain the President's certificate.

The sorting at the end handles out of order partial bundles, because the partialing occurs as the last bundle added but could be a lower percentage.