Game based allow/disallow cross corp abilities
I want to prevent players from using abilities they shouldn't be able to use. There are no companies in 18CO with an ability that can be used by other corporations, so I never want the option to show. Other games might have specific companies that can be used by third parties, and this function allows for that to be implemented.