Issue 2672 - No action indicator
During the stock round, all corporations are shown. This is good to have a general sense of relative positioning. However, many of the corporations will have no actions a player can take once they are selected. It would help reduce mental processing if there was a quick indication of whether there are actions available for that corporation.