Last/double cert funkiness
This cert must be bought last from wherever it is (corporate treasury or
market), issued last, and redeemed last. It can be sold by a player into
the market as they please. The current location of the double share is
marked with a 'd' on the corporatation charter.

Not included in this commit is the optional swap for the last cert by
the outgoing president in exchange for the presidency.