Add player choose to stock view
Make it possible to control from a step if choose should appear
in the view. It is useful in case there are several choose.

The main usage for this is in 18SJ where it is possible for a player
to trigger priority deal "stealing" due to an ability.

This commit contains the general code update needed to accomplish this.
Should not affect 1867 or 1817 which are the titles using choose so far.