Connections performance improvement
Change connections calculation to on-demand rather than when connected.

When a tile is layed, it walks the paths invalidating the cache of connected tiles
When route requests the connections at that point it calculates them if the cache is not already calculated.

This also simplifies the downgrade case as it invalidates it anyway

Stackprof goes from 138.76 seconds for 100 runs to 103.32 on 15528 (25% reduction)
The connect! goes from 33.92% of CPU to 16% (3.117% disconnect + 12.848% connect)
Loading unoptimized goes from 16.47 seconds to 14.23 seconds (15% reduction)

improves #2533