Make 18SJ (almost) feature complete
This more or less completes the implementation of 18SJ. There are
two things that are on draft, and which affect the common code but
I need to work more on those. (One is nationalization triggered by
phase change, and the other is discount for E train variant.)

The only common code change in this commit are addition of self
publisher as a publisher.