Use ability to active hex bonus for run
Removed buy special that was used before. This might break some
game but it can be removed as it is an alpha game and to much
effort to migrate.

Selecting company with the ability will now clear the route.
And unselecting will restore last route if such exists.

The selected company ability (when time: route) is passed in
route and is used when calculating revenue amount and revenue text.

Ability used is also included in RunRoute action. This is so
that ability can do use! when ability is activated.

Note! The hex bonus is not shown when historical route is used.
I think this is good, as you need to actively select the hex
bonus to apply it.