18SJ: Redesign main line tile lay/upgrade checks (fixes #3143)
Simplified main line checks, which should also speed up the game.

By using the icons in the tile hex the information if a hex is
a main line hex can be utilized to make a quicker decision.

For a main line hex, if tile lay/upgrade is a real improvement
we can now remove the icon, meaning the hex wont be checked
for main line improvement in the future.
And by just keeping track of number of improvements in each
main line we will know when a main line is completed.

By using process_tile_lay instead of process_action the code
was much simplified. The 18SJ track step now keeps most of
the knowledge of the tile lays' state.