18SJ: Handle priority deal chooser ability as a company (fixes #3144)
The PDC is handled as a company, with value 0, which is not part
of the initial auction, but instead given to the purchaser of Nils
Ericsson. It is handled as a no-buy company that is not counted
towards certificate limit, which only closes when the priority
deal choice ability is used.

This means that the "company" will appear in the player charter
and stay their for the rest of the game (until it is used) as
a reminder to the player.