Improve Round info on game card (fixes #3200)
Game keeps the current game info, and saves this the game is
stored.

The saved game info is used on game card. If game does not have
any it will show blank instead.

The round info is round name (minus Round) followed by the
or_description_short.
This means:
- 18MS show Operating 10 etc
- Standard 18xx show Operating 3.1 etc

The 18MS base class also contains a small refactoring which is
not connected to the fix.