Fix 18CO Prior Connected Paths Timing
In lay_tile, before check_track_restrictions! is called, the new tile is laid and the graph updated. This meant the original timing for calling prior_connected_paths was after the graph was updated, and thus we were examining newly connected paths as well. If the corporation could trace a loop back to the old segment of track, it would be returned as connected.

The new logic sets the prior_connected_paths before the graph is updated, so we are working with the true set of prior_connected_paths.