Changes after review comments plus some minor updates
General changes:
 - Restored purchasable_unsold_companies to step base class
   but let default be empty list.
 - Removed type attribute from minor base class. Handled by
   1824 subclassing minor instead. Added some extra attributes
   to the 1824 subclass to handle some more things.
 - Removed floatable attribute from corporation base class. Handled by
   1824 subclassing minor instead. Added some extra attributes
   to the 1824 subclass to handle some more things.
 - Restored float_str handling in general code, although have
   done some minor improvements (?)
 - Let finish_round in stock.rb be protected instead of private so
   it can be overridden by 1824.
 - Add "per share" in the payout log message.
 - Use same code for displaying train limit in Info page and for
   phase change. Previously phase change had an ugly printout in
   case it was a hash, but now both uses the prettified version.

For 1824 I have also found some small bugs when trying the game
out through phase 2 and 3 to make it playable. The major thing
that is missing is still possibility to exchange privates/minors/coal
for reserved and not reserved shares in staatsbahn and regionals.

One bug was that 3 trains did not appear as buyable as long as
1g train remained. In 1824 only the normal trains are considered
for phase changes so I have now subclassed depot to handle this
in 1824.