1824: Make Sarajevo offboard hex work functionally
Sarajevo is an offboard city, that use three hexes, with 6 termini.
Did not work with current implementation so am trying a different
approach. Now this uses 5 hexes instead of 3 so lanes is probably
a better solution, but this should be good enough as be playable.
Further improvements can be done in the future.