Add possibility to specify a selection of corporations for exchange
Change configuration attribute to corporations, as it now take an
array of corporation names.

Add a corporations(game) method to Exchange class, which returns
a list of all corporation (references) that can be candidates for
exchange. Hide the 'any' value handling in the Exchange ability
class - this return all corporations in the game.
Closed corporations are excluded.

Updated code to use the corporations(game) method. For 1860 and
1882 the game code only supports exactly one value in corporations
but it is assumed that the configuration for these games will
only include exactly one corporation name. For the other titles
the general changes should work even if they just specify one
corporation name or 'any', while the code now also supports
specifying more than one corporation name.