18SJ: Correct E train token handling (fixes #3173)
The SJ tokens is now made passable (neutral) when
route step is setup (if current entity owns E train)
and made impassible when dividend has been paid.

This might break some existing games that has broken
the rules of the game.
As this happens late in the games I think they can be
pinned if not completed, and archived if completed.