1824: Support exchange Mountain and Coal Railways
Mountain Railways is exchanged during SR in phase 3 by selecting
a regional, and click on Exchange. This exchange is for a 10% share.

Coal Railway is done as a Special Buy during SR in phase 3 or 4.
This exchange is for a 20% presidency share.
(The reason for separate ways is that the presidency share is set
to No Buy, and implementing support for doing normal exchange is
complicated. Think this way is good enough.)

When phase 5 is started the exchange of any remaining Coal Railways
is triggered.

Only change in general code is to show Special Buy during stock
round if any step responds to special_buy action.

Some minor changes of details (e.g. names) discovered during
play testing.