1824: Improve private rendering during first stock round
Removed BuyCompaniesAtFaceValue and unsold_purchasable_companies
as this was only used by 1824.

By setting owner to @bank for all companies, the render_bank_companies
will render the privates in a better way. There are some minor
differences (button text and location) but they are OK. I suppose it
is good to have consistency between titles.

Hide bought coal companies, and corporations that cannot be bought,
during SR 1.

Added corporations attribute to bank. This is needed when buying
companies that bank own. (Which is what is required to get stock
view to work in 1824.) Now that is some 18xx (e.g. 1825) that allow
companies to be sold to bank so it might be useful for other titles
as well.