1824 mainly: Make Coal Railways minors
The first attempt was to make Coal Railways corporations but they function
in all ways as a minor so this commits makes them minors.

The main reason for having them as corporations was that these had
5 different purchase prices (120G - 200G) so using the parring worked.

But it turned out to cause problems down the line so I now go back
to having them as minors.

Now, it still needs to support buying for different prices. And I have
now added support for having multiple prices by adding an interval
attribute to companies. In case this is set (default nil) the stock
view will render buy buttons similar to par buttons, when buying
private from bank.
In case interval is not specified as attribute this change has no
effect.