1824: Add an interrupt stock exchange step when 4 trains bought
In 1824 when the first 4 train is bought any remaining Mountain
Railways' companies need to be exchanged for shares in one of
the regionals. (Often multiple corporations possible to choose from.)

This is handled by having a stock view with only Exchange possible
during the OR.

This also means that game_page.rb was changed to allow for stock
market rendering during an OR (in case buy_shares is available).