Show double certificates as d
Generalized the 1849 solution. Added a double_cert attribute to
display 'd'. While keeping the last_cert for the specific 1849
functionality.

Double certs are used in 1893 as a simple 20% certificate.
In the physical game the 20% share is the 2nd certificate (with
Director certificate as number one) but as certificate 1-3 all
are reserved, and this has no other game effect it probably does
not matter. Putting the 20% made the number of shares decrease to 5.