Link to BoardGameGeek in the Game Info section of the Info tab
Put the BoardGameGeek link next to the existing “more info” link to the wiki.

If we don’t have a link to the wiki, linking to the Google front page seems unhelpful, so let’s not show the link in that case.

Since GAME_INFO_URL will always point to the wiki, make that clearer by renaming it to GAME_WIKI_URL.