1893: Complete Starting Package
Starting Package is all privates and the companies that corresponds to
minors. All (but one) must be sold before anything else can be bought.

The game starts with the drafting. In case all but one is sold, the
game continues with SR1. In case all passes before that, the games
continues with OR1.1.

After OR1.2, there is a special auction round, if any of starting
package remains, where certificates from the starting package is
put up for Dutch auction, until all is sold, and SR2 commences.
In this Dutch auction price is decreased by 10 until it reaches
50% of starting when some player is forced to buy it.

(Note: Max 2 of the minors can be up for sale at any times.)