1893: Complete merge actions
Added Merge step for AGV and HGK. This will either do an auto
merge if late phase, and/or give the possibility to vote if merge
should take place or not.

If merge - minors + private (in case of HGK) are merged into
AGV/HGK which floats. Resources moved to AGV/HGK. Reserved
shares are exchanged.

Added a discard trains step which is used as the director of
the floated merged corporation may discard any train.

Potential Merger round takes place betwee Stock Round and Operating
Round as well as between first OR and second OR in each turn.