add sort argument to shares_by_corporation, use during bankruptcy
If sort is true, then the hash's keys are sorted so that when the caller
iterates through the corporations, they iterate in operating order. This is
relevant for games that do not end in bankruptcy, where the order the
bankrupting player sells the shares can impact future operating order.

Technically the current site behavior for 1817 does not violate the rules, but
as #4181 describes, the return value of `shares_by_corporation` can potentially
be reordered by a seemingly unrelated code change. Getting a well defined order
in place for selling these shares avoids that being a problem in the future.

This breaks breaks about 6% of 1817-family games. No other games are broken,
though many will see a change in the logged order of shares sold during
bankruptcy.

Replace the 1817NA spec; 27066 is among those broken by this change