Make it possible to override orientation check
This is needed for 1893, when upgrading Leverkusen hex, from
yellow to green.

Yellow Leverkusen hex has two towns, each on a separate track.
The green Leverkusen hex hex have track at same edges, but the
track makes an X with a city in the middle. It does remove the
towns.