Make it possible to draft minors for a min_bid
Some games (e.g. 1893) has the possibility to draft a minor for
its "face value". Have added a may_draft?(minor) method that if
this return true, then a "Buy for XX" button appear during auction/draft.
The XX is the amount that is returned by min_bid(minor).
Buying the minor will result in a bid action with minor set,
and price the min_bid.

Default is that no such button appear. The step need to implement
may_draft?(minor) and have it return true for it to appear.