1893: Allow city upgrade even if no new track is usable (fixes #5836)
The bug was when upgrading Köln from brown to gray - which caused a
"must use new track", but gray Köln has no new track. The rules
are not explicit if it is allowed to upgrade a city tile where new
track is not reachable but my interpretation is that this should
be allowed.