Change minor check in buy trains actions (fixes #5843)
By making the check in a method it is easy for titles to override
method if they want to allow a minor to be able to buy trains.

If instead an override of actions' method it might cause a bug
if code is forgotten - which is what happended in 1893 (and 1824).

A quick scan through of existing titles did not find any other
title where minors can buy trains so I have not updated any
other title.