[1893] Support auto buy until float (fixes #5794)
As shares in 1893 are moved from IPO to market when parred,
we need to change the check in general code that checks if auto
order was for market or IPO. "Buy until float" has IPO as source.
So the check is extracted to a method, which 1893 can override,
so that "auto buy until float" always buys from market.