[1893] Add possibility to purchase EVA from other player
This commit makes it possible for the active player to purchase
the EVA company (for any amount - upto cash limit) at two times:
- During the stock round (is not counted as an action)
- During train purchase step

The purchaser is the active player, not the minor/corporation.
The purchaser pay.

Note! The rules does allow for buy/sell at any time, and this
would also be regarded as an action, for determination of first
player, but this implementation is a simplication that should
be good enough to at least give the players an option to buy EVA.