[18Rhl] Add logos for all corporations, add icons
S (Steel) and K (Coal) is used as both icons and logos (tokens).
ERh and RGE is only used as icons.
The remaining are for the corporations in the game.

Made some corrections on the map.

Implemented the variable placement of one S and one K icon on the
map depending on random number.