[18Rhl] Disallow double visits to yellow Rhine Metropolis
When Duisburg, Düsseldorf and Köln are yellow hexes, they have three
separate city slots. Two of these are on one side of Rhine, and the
third is on the other. It is not allowed to include the two city
slots on one side of Rhine in the same route (as it is the same city).
Having a route from one side of Rhine to the other side of Rhine
and visiting the different sides of the same hex, is allowed though.

When the Rhine Metropolis hex is upgraded to green (or later) the
hex will be a normal "one city - three slots" and then the normal
game logic will stop the hex being visited twice in one route.

This is a 2nd attempt to implement this feature - the 1st attempt
was reverted, as it need to be reviewed, and the 1st attempt also
did not work as it made routes through green Rhine metropolis cities
illegal.