[18Rhl] Complete implementation of all company abilities
Some minor changes was needed in general code to cover the case when
a player owned company performs a tokening - as the tokener is
then not the comany, or company owner, but the current entity
(corporation) that is operating. This should not affect existing
titles, as a check is done if owner is player.