[1893] Do not allow buy of company when cert limit reached
This probably only applies when trying to buy an AdSK bond late
in the game.

This might break games as it was previously possible to buy
AdSK bonds to exceed cert limit (although those had to be sold
to be allowed to pass in SR), so any games this break can be
pinned (or archived if finished).