18EU Minor Companies Merge Timing
Minor Companies may never merge with a Corporation once it has operated. However, if a Corporation is started but not floated, any Minor Companies that would be legally allowed to merge are allowed to do so in a subsequent Stock Round just as if the Corporation had been started that Stock Round.