Make home page navigation preserve unrelated query params.
This is needed to support filtering by e.g. game title, since otherwise the filters will be dropped as soon as you paginate.

The immediate effect is that paginating in one section won't reset the other sections to page 0.

Introduces a class in Lib::Params for stateful manipulation of search/query params.