Check for auto-buy completion before checking available actions.
Auto-buy programs would not mark themselves as complete if buy_shares was not an available action (a common scenario when spending all your money to float). Formerly only a nuisance, but with the option to transition to auto-pass it actually breaks that functionality. Easy fix, just check for completion before checking if buying is allowed.

This *will* change behavior for buy X shares programs in buy-sell titles. They would previously generate a pass action after the last buy and terminate on the next turn. Now they will terminate after the last buy and not generate a pass (unless conversion to auto-pass is enabled). This seems like an improvement since the option to auto-pass exists, and the default behavior has been aligned with buy until float.

I've fixed the one fixture that was affected by this. My understanding is that real games won't need patching because the generated pass action remains valid even if the code would no longer generate it.