[1846] Fix hotseat draft with real player names.
The draft is hidden by default and requires master mode to show details for non-acting players. This is not required for logged-out users and users not in the game, which also makes it work in hotseat most of the time. However, in a hotseat game where one of the players has the same name as the logged-in user, the code gets confused and requires master mode.