18ChEv Render company specific choices
In the stock screen, this now only renders choices for a choose action
for a specific company if that company is selected. It also now calls
'entity_choices' (if it is available) instead of 'choices' (which it calls
if it is not) to permit having different choices for different entities.

Formerly, choice showed regardless if the stock was selected or not.
This puts it in with buy/sell unfolding. In this case, it was used to
support the 18ChEv share gifting buttons. I haven't found anywhere
else that per company options are used in the stock screen that will
be effected by this change